% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-package.R
\docType{package}
\name{notebook-package}
\alias{notebook}
\alias{notebook-package}
\title{notebook: Notebook templates for ANPC analysis and sample preparation}
\description{
Provides templates for notebook for analysis at the Australian National Phenome Centre
}
\author{
\strong{Maintainer}: Julien Wist \email{julien.wist@murdoch.edu.au} (\href{https://orcid.org/0000-0002-3416-2572}{ORCID})

}
\keyword{internal}
